Separated interface and implementation for sample_websocket_server
