+++ src/components/transport_manager/test/CMakeLists.txt
@@ -49,6 +49,7 @@ set(EXCLUDE_PATHS
 if (NOT BUILD_CLOUD_APP_SUPPORT)
   list(APPEND EXCLUDE_PATHS
     ${CMAKE_CURRENT_SOURCE_DIR}/websocket_connection_test.cc
+    ${CMAKE_CURRENT_SOURCE_DIR}/sample_websocket_server.cc
   )
 endif()
